refactor: view model layer updates
