+++ app/src/main/java/com/example/partymaker/viewmodel/auth/IntroViewModel.java
@@ -1,9 +1,11 @@
-package com.example.partymaker.viewmodel;
+package com.example.partymaker.viewmodel.auth;
 
 import android.app.Application;
 import androidx.annotation.NonNull;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
+
+import com.example.partymaker.viewmodel.BaseViewModel;
 
 /** ViewModel for IntroActivity handling intro/onboarding state */
 public class IntroViewModel extends BaseViewModel {
