+++ app/src/main/java/com/example/partymaker/viewmodel/features/GptViewModel.java
@@ -1,4 +1,4 @@
-package com.example.partymaker.viewmodel;
+package com.example.partymaker.viewmodel.features;
 
 import android.app.Application;
 import androidx.annotation.NonNull;
@@ -8,6 +8,8 @@ import com.example.partymaker.data.api.OpenAiApi;
 import com.example.partymaker.data.model.ChatMessage;
 import com.example.partymaker.utils.security.core.SecureConfigManager;
 import com.example.partymaker.utils.infrastructure.system.ThreadUtils;
+import com.example.partymaker.viewmodel.BaseViewModel;
+
 import java.util.ArrayList;
 import java.util.List;
