+++ app/src/main/java/com/example/partymaker/viewmodel/groups/SimplifiedGroupViewModel.java
@@ -1,4 +1,4 @@
-package com.example.partymaker.viewmodel;
+package com.example.partymaker.viewmodel.groups;
 
 import android.app.Application;
 import androidx.annotation.NonNull;
@@ -6,6 +6,8 @@ import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import com.example.partymaker.data.model.Group;
 import com.example.partymaker.data.model.User;
+import com.example.partymaker.viewmodel.BaseViewModel;
+
 import java.util.ArrayList;
 import java.util.List;
