+++ app/src/main/res/layout/activity_party_change_date.xml
@@ -8,7 +8,7 @@
     android:fitsSystemWindows="true"
     android:orientation="vertical"
     android:padding="20dp"
-    tools:context="com.example.partymaker.ui.features.groups.ChangeDateActivity">
+    tools:context="com.example.partymaker.ui.features.groups.managment.ChangeDateActivity">
 
     <ImageView
         android:layout_width="160dp"
